Change hitting style
This commit changes the hitting style so that the play can now only hit the ball if they are hovering of it. Additionally, the direction the player hits the ball has been reversed, meaning that the player now has to drag to where they want the ball to go, rather than dragging backwards (e.g., Angry Birds style).