Add gorilla with dumpy, and fix some physics
Trying to make the gorilla's feet not effect the ball and have the ball also be able to sit on the gorilla's back. We increased the amount of physics checks made by unity to see if it fixed movement Jitters and added some more conditionals to help with differing between wall jumping and ground jumping.
Co-authored-by: burge337 <burge337@morris.umn.edu>