Fixed wall jumping issue by implementing new wall jump
Added sensors to the Gorilla to detect when a wall jump is needed, due to the tilemap only having one tag. Not a perfect solution, because as of right now we cannot wall jump off of tiny walls. Also both versions of resetting the gorillas jump on contact with "ground" is implemented, so we get a weird jump reset on the side of any part of the tilemap