Fix ball power/trajectory issue
This commit fixes the issue where the player  could drag their mouse as far as they could and max out the power. When this happened, it created this weird problem where the ball would launch at a 45 degree angle. This commit changes that so the endPoint of the trajectory line is used in the power calculations rather than the mouse input point.