Fix issue where the ball loses all forward momentum on collision
This commit fixes the issue where the ball loses all of its forward momentum on a collision. I believe the main cause of the problem was due to the z-axis of rigid body of the ball being constrained. When the z-axis is constrained, the ball doesn't have angular velocity or inertia.