Fix weird ball behaviour when trying to hit it while it is moving
This commit fixes the weird behaviour that would occur when trying to hit the ball while it is moving. To fix this issue, this commit makes so that the ball's velocity is set 0 right before the new force is applied. Additionally, part of the problem was that the script was still using the mouse pointer as the start position rather than the ball's position.