Fix ball physics being weird when the ball is moving
Previously we were using mouse position from the 'OnMouseButtonDown' event which meant that if you held the button down while the ball was moving the position would quickly become out of date. This has been changed to take place in the 'OnMouseButtonUp' event, using the ball's position as the starting point instead.