Tutorial Walls
Tutorial Summon script colliders have an input to activate an object now and I added some of them in to the tutorial level. Also some jank is happening with the GorillaBuffer. Current State: Because Gorilla is relying on the downward box cast to be able to press the jump button, he can't jump when he is on the complete edge of his collider. Potential solution: Find out how to widen the boxRay