Write comments for gorillaController / Fix some input buffer issues
Wrote comments for all of the things we are doing inside of the methods in gorillaControler. Then I made the downRay wider so the gorilla can still jump when standing at the edge of his hitbox. Also, I made it so all the tutorialUI is set to deactive instead of active.