Change space level 2 to fix issue where the player could lose the ball
This commit fixes an issue where the player could lose the ball during the camera change. The ball could bounce back through the opening  and the player would not be able to go back and get it. This commit changes that by changing the position of the camera change and adding a new one. The new camera changes make it possible to go back to the ball if it goes back through the opening.