Improve gorilla back ball traction
I've given the back of the gorilla 100 friction and added an 'OnCollisionEnter2D' call that freezes the z rotation of the ball when hitting the gorilla. This makes the ball bounce a little less sporadically off of the gorilla. It also allows for easy horizontal movement while having carrying the ball. Currently only implemented in JungleLevel3.