Add casual leaderboard
This makes some of the leaderboard functions in PlayfabManager a bit more flexible, but currently there is a hard-coded reference to spacelevel3, so if we swap around which scene is the final level that will break. I'm assuming we are wrapping up the project and won't be making more levels however.